third_party/blink/web_tests/fast/js/mozilla/strict/10.4.2-expected.txt

PASS 'use strict'; eval('010') threw exception of type SyntaxError.
PASS true === true
PASS 'use strict'; eval('"use strict"; 010') threw exception of type SyntaxError.
PASS eval('"use strict"; 010') threw exception of type SyntaxError.
PASS true === true
PASS true === true
PASS true === true
PASS Function('"use strict"; 010') threw exception of type SyntaxError.
PASS true === true
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
